requirements.txt

pandas
numpy
notebook
scikit-learn
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML



